requirements.txt
altair==5.0.1
appdirs==1.4.4
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
biopython==1.81
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cssselect==1.2.0
decorator==5.1.1
fake-useragent==1.1.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
Jinja2==3.1.2
jsonschema==4.17.3
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.24.3
packaging==23.1
pandas==2.0.2
parse==1.19.0
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
protobuf==4.23.2
pyarrow==12.0.0
pydeck==0.8.1b0
pyee==8.2.2
Pygments==2.15.1
Pympler==1.0.1
pyppeteer==1.0.2
pyquery==2.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.31.0
requests-html==0.10.0
rich==13.4.2
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
streamlit==1.23.1
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing-extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.16
validators==0.20.0
w3lib==2.1.1
watchdog==3.0.0
websockets==10.4
zipp==3.15.0
